Integrated gallery component.
