+++ package.json
@@ -48,6 +48,7 @@
     "boostrap": "^2.0.0",
     "konva": "^8.3.2",
     "moment": "^2.29.1",
+    "ng-gallery": "^6.0.0",
     "path": "^0.12.7",
     "prisma": "^3.7.0",
     "rxjs": "~7.4.0",
