+++ src/app/app.component.html
@@ -1,1 +1,1 @@
-<app-virtual-canvas></app-virtual-canvas>
+<router-outlet></router-outlet>
