+++ src/app/app.module.ts
@@ -2,9 +2,11 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { ScrollingModule } from '@angular/cdk/scrolling';
+import { GalleryModule } from 'ng-gallery';
 import { AppComponent } from './app.component';
 import { components } from './components';
 import { pipes } from './pipes';
+import { AppRoutingModule } from './app-routing.module';
 
 @NgModule({
   declarations: [
@@ -15,7 +17,9 @@ import { pipes } from './pipes';
   imports: [
     BrowserModule,
     BrowserAnimationsModule,
-    ScrollingModule
+    ScrollingModule,
+    GalleryModule,
+    AppRoutingModule
   ],
   providers: [],
   bootstrap: [AppComponent]
