+++ src/app/components/image-viewer/image-viewer.component.html
@@ -1,0 +1,3 @@
+<div *ngIf="items$ | async as items" class="gallery-container">
+  <gallery [items]="items" thumbPosition="bottom"></gallery>
+</div>
