+++ src/app/components/image-viewer/image-viewer.component.scss
@@ -1,0 +1,4 @@
+.gallery-container,
+gallery {
+  height: 100%;
+}
