+++ src/app/components/index.ts
@@ -1,5 +1,7 @@
 import { VirtualCanvasComponent } from './virtual-canvas/virtual-canvas.component';
+import { ImageViewerComponent } from './image-viewer/image-viewer.component';
 
 export * from './virtual-canvas/virtual-canvas.component';
+export * from './image-viewer/image-viewer.component';
 
-export const components = [VirtualCanvasComponent];
+export const components = [VirtualCanvasComponent, ImageViewerComponent];
